- Moved search bar to own file - Introduced thread-version of messageslist etc - Refactored TimelineItemVariant, lifted reply to details and thread details - Refactored MessageDrawerContent to use MessageDrawerContentInterface using generics, so that the actions can be implemented at the call site. Used differently for Room and Thread list -
This commit makes some unpleasant design choices, but they are easy to fix in coming refactors
