+++ app/src/main/java/com/app/radiator/ui/components/MessageComposerState.kt
@@ -173,7 +173,7 @@ class MessageComposerState(
         this.msgContent.value, act.eventId
       )
     }
-    timeline.send(timelineAction)
+    timeline.timelineSend(timelineAction)
     setState(ComposerState.NewMessage)
   }
 }
