+++ app/src/main/java/com/app/radiator/ui/routes/Room.kt
@@ -12,17 +12,21 @@ import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.ModalBottomSheetLayout
 import androidx.compose.material.ModalBottomSheetState
 import androidx.compose.material.ModalBottomSheetValue
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Close
+import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material.icons.filled.Edit
+import androidx.compose.material.icons.filled.Favorite
 import androidx.compose.material.icons.filled.KeyboardArrowDown
 import androidx.compose.material.icons.filled.MoreVert
+import androidx.compose.material.icons.filled.Refresh
 import androidx.compose.material.icons.filled.Search
+import androidx.compose.material.icons.filled.Send
+import androidx.compose.material.icons.filled.Share
+import androidx.compose.material.icons.filled.ThumbUp
 import androidx.compose.material.icons.outlined.Person
 import androidx.compose.material.icons.outlined.Settings
 import androidx.compose.material.rememberModalBottomSheetState
@@ -34,10 +38,8 @@ import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Immutable
@@ -49,18 +51,11 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
-import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.focus.focusTarget
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalSoftwareKeyboardController
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
@@ -78,13 +73,15 @@ import com.app.radiator.matrix.timeline.ProfileDetails
 import com.app.radiator.matrix.timeline.TimelineState
 import com.app.radiator.ui.components.Avatar
 import com.app.radiator.ui.components.LoadingAnimation
-import com.app.radiator.ui.components.MessageDrawerActionType
 import com.app.radiator.ui.components.MessageDrawerContent
 import com.app.radiator.ui.components.MessageComposer
 import com.app.radiator.ui.components.MessageComposerState
 import com.app.radiator.ui.components.ComposerState
+import com.app.radiator.ui.components.MessageDrawerAction
+import com.app.radiator.ui.components.MessageDrawerContentInterface
+import com.app.radiator.ui.components.SearchBar
 import com.app.radiator.ui.components.general.CenteredRow
-import com.app.radiator.ui.components.showAddOnToMessage
+import kotlinx.collections.immutable.persistentListOf
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
@@ -136,55 +133,6 @@ sealed class DropDownMenuItems(
 @Composable
 fun PreviewSearchBar() {
   SearchBar(onSearch = {}, onDone = {})
-}
-
-@OptIn(ExperimentalComposeUiApi::class)
-@Composable
-fun SearchBar(onSearch: (String) -> Unit, onDone: () -> Unit) {
-  val (messageInput, setMessage) = remember { mutableStateOf("") }
-  val keyboardController = LocalSoftwareKeyboardController.current
-  val focusRequester = remember { FocusRequester() }
-  Row(
-    modifier = Modifier.background(color = Color.White),
-    verticalAlignment = Alignment.CenterVertically
-  ) {
-    OutlinedTextField(
-      modifier = Modifier
-        .fillMaxWidth()
-        .focusTarget()
-        .weight(0.5f)
-        .focusRequester(focusRequester),
-      value = messageInput,
-      colors = TextFieldDefaults.colors(
-        focusedContainerColor = Color.White,
-        unfocusedContainerColor = Color.White
-      ),
-      textStyle = TextStyle(fontSize = 14.sp),
-      onValueChange = setMessage,
-      keyboardOptions = KeyboardOptions(imeAction = ImeAction.Send),
-      keyboardActions = KeyboardActions(
-        onSend = {
-          onSearch(messageInput)
-        }, onDone = {
-          keyboardController?.hide()
-          onDone()
-        }, onNext = {
-
-        }),
-      label = { Text("Search for...") },
-      trailingIcon = {
-        Icon(
-          imageVector = Icons.Default.Close,
-          contentDescription = "Exit search mode",
-          modifier = Modifier.clickable {
-            onDone()
-          })
-      }
-    )
-  }
-  LaunchedEffect(Unit) {
-    focusRequester.requestFocus()
-  }
 }
 
 @Preview
@@ -303,63 +251,6 @@ fun RoomTopBar(
   }
 }
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ThreadTopBar(
-  avatarData: AvatarData,
-  onSearch: (String) -> Unit,
-) {
-
-  // val interactionSource = remember { MutableInteractionSource() }
-  // val (expanded, setExpanded) = remember { mutableStateOf(false) }
-  val (searchingRoom, setSearchRoom) = remember {
-    mutableStateOf(false)
-  }
-  fun cancelSearch() {
-    setSearchRoom(false)
-  }
-
-  @Composable
-  fun showSearch() {
-    IconButton(onClick = {
-      setSearchRoom(true)
-    }) {
-      Icon(imageVector = Icons.Default.Search, contentDescription = "Search in room for...")
-    }
-  }
-
-  Box(modifier = Modifier.border(width = 2.dp, color = Color.LightGray)) {
-    TopAppBar(
-      title = {
-        Row(
-          verticalAlignment = Alignment.CenterVertically,
-          horizontalArrangement = Arrangement.Start,
-        ) {
-          Avatar(modifier = Modifier, avatarData = avatarData)
-          Spacer(Modifier.width(10.dp))
-          Text(text = avatarData.name!!)
-          Spacer(Modifier.weight(1.0f))
-        }
-      },
-      actions = {
-        if (!searchingRoom) {
-          showSearch()
-        } else {
-          SearchBar(onSearch = onSearch, onDone = { cancelSearch() })
-        }
-        IconButton(onClick = {
-          // setExpanded(true)
-        }) {
-          Icon(
-            imageVector = Icons.Default.MoreVert,
-            contentDescription = "Open Options"
-          )
-        }
-      }
-    )
-  }
-}
-
 @OptIn(ExperimentalMaterialApi::class)
 @Composable
 fun RoomRoute(
@@ -392,21 +283,16 @@ fun RoomRoute(
     modifier = Modifier,
     sheetState = itemActionsBottomSheetState,
     sheetContent = {
-      MessageDrawerContent { event ->
+      MessageDrawerContent(TimelineMessageDrawer) { event ->
         coroutineScope.launch {
           itemActionsBottomSheetState.hide()
           val item = clickedItem.value!!.copy(reactions = listOf())
           when (event) {
-            MessageDrawerActionType.Reply -> messageComposer.setState(ComposerState.Reply(item))
-            MessageDrawerActionType.ThreadReply -> messageComposer.setState(
-              ComposerState.ThreadReply(
-                item
-              )
-            )
-
-            MessageDrawerActionType.React -> messageComposer.setState(ComposerState.React(item))
-            MessageDrawerActionType.Edit -> messageComposer.setState(ComposerState.Edit(item))
-            MessageDrawerActionType.Delete, MessageDrawerActionType.Share, MessageDrawerActionType.Quote, MessageDrawerActionType.NewMessage -> {}
+            TimelineMessageDrawerActionType.Reply -> messageComposer.setState(ComposerState.Reply(item))
+            TimelineMessageDrawerActionType.ThreadReply -> messageComposer.setState(ComposerState.ThreadReply(item))
+            TimelineMessageDrawerActionType.React -> messageComposer.setState(ComposerState.React(item))
+            TimelineMessageDrawerActionType.Edit -> messageComposer.setState(ComposerState.Edit(item))
+            TimelineMessageDrawerActionType.Delete, TimelineMessageDrawerActionType.Share, TimelineMessageDrawerActionType.Quote -> {}
           }
         }
       }
@@ -483,118 +369,6 @@ fun RoomRoute(
 
 @OptIn(ExperimentalMaterialApi::class)
 @Composable
-fun ThreadRoute(
-  navController: NavHostController,
-  timelineState: TimelineState,
-  messageComposer: MessageComposerState,
-  threadEventId: String,
-) {
-  val lazyListState = rememberLazyListState()
-  val coroutineScope = rememberCoroutineScope()
-  val messages = timelineState.threadItemFlow(threadEventId).collectAsState()
-  val lastSearchHitIndex = remember { mutableStateOf(0) }
-
-  val itemActionsBottomSheetState = rememberModalBottomSheetState(
-    initialValue = ModalBottomSheetValue.Hidden,
-  )
-
-  val clickedItem: MutableState<TimelineItemVariant.Event?> = remember {
-    mutableStateOf(null)
-  }
-
-  LaunchedEffect(itemActionsBottomSheetState.isVisible) {
-    if (!itemActionsBottomSheetState.isVisible) {
-      clickedItem.value = null
-    }
-  }
-
-  ModalBottomSheetLayout(
-    modifier = Modifier,
-    sheetState = itemActionsBottomSheetState,
-    sheetContent = {
-      MessageDrawerContent { event ->
-        coroutineScope.launch {
-          itemActionsBottomSheetState.hide()
-          val item = clickedItem.value!!.copy(reactions = listOf())
-          when (event) {
-            MessageDrawerActionType.Reply -> messageComposer.setState(ComposerState.Reply(item))
-            MessageDrawerActionType.ThreadReply -> messageComposer.setState(
-              ComposerState.ThreadReply(
-                item
-              )
-            )
-
-            MessageDrawerActionType.React -> messageComposer.setState(ComposerState.React(item))
-            MessageDrawerActionType.Edit -> messageComposer.setState(ComposerState.Edit(item))
-            MessageDrawerActionType.Delete, MessageDrawerActionType.Share, MessageDrawerActionType.Quote, MessageDrawerActionType.NewMessage -> {}
-          }
-        }
-      }
-    }
-  ) {
-    Box(modifier = Modifier.background(color = Color.White)) {
-      Scaffold(content = { padding ->
-        ThreadList(
-          navController,
-          timelineState = timelineState,
-          padding = padding,
-          lazyListState = lazyListState,
-          messages = messages,
-          itemActionsBottomSheetState = itemActionsBottomSheetState,
-          clickedItemPublisher = clickedItem,
-          requestMore = { timelineState.requestMore() }
-        )
-      }, bottomBar = {
-        MessageComposer(messageComposer)
-      }, topBar = {
-        ThreadTopBar(
-          avatarData = timelineState.avatar(),
-          onSearch = { searchString ->
-            coroutineScope.launch {
-              searchTimeline(searchString, lazyListState, lastSearchHitIndex, messages)
-            }
-          },
-        )
-      }, floatingActionButton = {
-        // TODO: _maybe_ have to toggle off when at the bottom, but this comes with an additoinal cost
-        //  it means this composable will get a recomposition triggered _every time the user scrolls_
-        //  - I'm not willing to pay that cost, right now
-        FloatingActionButton(
-          onClick = {
-            coroutineScope.launch {
-              if (messages.value.size - lazyListState.firstVisibleItemIndex > 25) {
-                lazyListState.scrollToItem(messages.value.size)
-              } else {
-                lazyListState.animateScrollToItem(messages.value.size)
-              }
-            }
-          },
-          shape = CircleShape,
-          modifier = Modifier
-            .align(Alignment.BottomCenter)
-            .size(40.dp),
-          containerColor = MaterialTheme.colorScheme.surfaceVariant,
-          contentColor = MaterialTheme.colorScheme.onSurfaceVariant
-        ) {
-          Icon(Icons.Default.KeyboardArrowDown, "")
-        }
-      }, floatingActionButtonPosition = FabPosition.End
-      )
-
-      // Auto-scroll when new timeline items appear
-      LaunchedEffect(messages.value) {
-        coroutineScope.launch {
-          if (lazyListState.isScrolledToTheEnd() && !lazyListState.isScrollInProgress) {
-            lazyListState.animateScrollToItem(messages.value.size)
-          }
-        }
-      }
-    }
-  }
-}
-
-@OptIn(ExperimentalMaterialApi::class)
-@Composable
 fun MessageList(
   navController: NavHostController,
   timelineState: TimelineState,
@@ -626,7 +400,7 @@ fun MessageList(
     ) { index, timelineItem ->
       when (timelineItem) {
         is TimelineItemVariant.Event -> {
-          if (timelineItem.userCanSee && timelineItem.threadId == null) {
+          if (timelineItem.userCanSee && timelineItem.threadDetails == null) {
             RoomMessageItem(
               item = timelineItem,
               onClick = {
@@ -635,125 +409,6 @@ fun MessageList(
                 } else {
                   Log.d("RoomMessageItemClick", "Clicked message item $timelineItem.eventId")
                 }
-              },
-              onClickHold = {
-                coroutineScope.launch {
-                  itemActionsBottomSheetState.show()
-                  clickedItemPublisher.value = it
-                }
-              },
-            )
-          }
-        }
-
-        is TimelineItemVariant.Virtual -> VirtualItem(timelineItem = timelineItem)
-        TimelineItemVariant.Unknown -> {}
-        is TimelineItemVariant.Fill -> {}
-      }
-
-      if (reachedTopOfList(index)) {
-        requestMore()
-      }
-    }
-  }
-}
-
-@OptIn(ExperimentalMaterialApi::class)
-@Composable
-fun ThreadList(
-  navController: NavHostController,
-  timelineState: TimelineState,
-  padding: PaddingValues,
-  lazyListState: LazyListState,
-  messages: State<List<TimelineItemVariant>>,
-  itemActionsBottomSheetState: ModalBottomSheetState,
-  clickedItemPublisher: MutableState<TimelineItemVariant.Event?>,
-  requestMore: () -> Unit,
-) {
-  fun reachedTopOfList(index: Int): Boolean {
-    return index == 0
-  }
-
-  val coroutineScope = rememberCoroutineScope()
-  LazyColumn(
-    modifier = Modifier
-      .fillMaxSize()
-      .padding(padding),
-    state = lazyListState,
-    horizontalAlignment = Alignment.Start,
-    verticalArrangement = Arrangement.Bottom,
-    reverseLayout = false
-  ) {
-    itemsIndexed(
-      items = messages.value,
-      contentType = { _, timelineItem -> timelineItem.contentType() },
-      key = { _, timelineItem -> timelineItem.id() },
-    ) { index, timelineItem ->
-      when (timelineItem) {
-        is TimelineItemVariant.Event -> {
-          RoomMessageItem(
-            item = timelineItem,
-            onClick = {
-              Log.d("RoomMessageItemClick", "Clicked message item $timelineItem.eventId")
-            },
-            onClickHold = {
-              coroutineScope.launch {
-                itemActionsBottomSheetState.show()
-                clickedItemPublisher.value = it
-              }
-            },
-          )
-        }
-
-        is TimelineItemVariant.Virtual -> VirtualItem(timelineItem = timelineItem)
-        TimelineItemVariant.Unknown -> {}
-        is TimelineItemVariant.Fill -> {}
-      }
-
-      if (reachedTopOfList(index)) {
-        requestMore()
-      }
-    }
-  }
-}
-
-@OptIn(ExperimentalMaterialApi::class)
-@Composable
-fun MessageList(
-  padding: PaddingValues,
-  lazyListState: LazyListState,
-  messages: State<List<TimelineItemVariant>>,
-  itemActionsBottomSheetState: ModalBottomSheetState,
-  clickedItemPublisher: MutableState<TimelineItemVariant.Event?>,
-  requestMore: () -> Unit,
-  threadId: String?,
-) {
-  fun reachedTopOfList(index: Int): Boolean {
-    return index == 0
-  }
-
-  val coroutineScope = rememberCoroutineScope()
-  LazyColumn(
-    modifier = Modifier
-      .fillMaxSize()
-      .padding(padding),
-    state = lazyListState,
-    horizontalAlignment = Alignment.Start,
-    verticalArrangement = Arrangement.Bottom,
-    reverseLayout = false
-  ) {
-    itemsIndexed(
-      items = messages.value,
-      contentType = { _, timelineItem -> timelineItem.contentType() },
-      key = { _, timelineItem -> timelineItem.id() },
-    ) { index, timelineItem ->
-      when (timelineItem) {
-        is TimelineItemVariant.Event -> {
-          if (timelineItem.userCanSee) {
-            RoomMessageItem(
-              item = timelineItem,
-              onClick = {
-                Log.d("RoomMessageItemClick", "Clicked message item")
               },
               onClickHold = {
                 coroutineScope.launch {
@@ -836,3 +491,25 @@ fun VirtualItem(timelineItem: TimelineItemVariant.Virtual) {
 // we have to compare (less-than) against 3, to "be sure" that we're at the end. Seems lazy column isn't *exact* in it's measurements
 fun LazyListState.isScrolledToTheEnd() =
   (layoutInfo.totalItemsCount - (layoutInfo.visibleItemsInfo.lastOrNull()?.index ?: 0)) < 3
+
+@Immutable
+enum class TimelineMessageDrawerActionType {
+  Reply, ThreadReply, React, Edit, Delete, Share, Quote
+}
+
+object TimelineMessageDrawer : MessageDrawerContentInterface<TimelineMessageDrawerActionType> {
+  private val actions = persistentListOf(
+    MessageDrawerAction("Reply", Icons.Default.Send, desc = "Reply to message", action = TimelineMessageDrawerActionType.Reply),
+    MessageDrawerAction(
+      "Reply in thread", Icons.Default.Refresh, desc = "Reply in thread", action = TimelineMessageDrawerActionType.ThreadReply
+    ),
+    MessageDrawerAction(
+      "Reaction", Icons.Default.ThumbUp, desc = "React to message", action = TimelineMessageDrawerActionType.React
+    ),
+    MessageDrawerAction("Edit", Icons.Default.Edit, desc = "Edit message", action = TimelineMessageDrawerActionType.Edit),
+    MessageDrawerAction("Delete", Icons.Default.Delete, desc = "Delete message", action = TimelineMessageDrawerActionType.Delete),
+    MessageDrawerAction("Share", Icons.Default.Share, desc = "Share", action = TimelineMessageDrawerActionType.Share),
+    MessageDrawerAction("Quote", Icons.Default.Favorite, desc = "Quote", action = TimelineMessageDrawerActionType.Quote),
+  )
+  override fun actions(): List<MessageDrawerAction<TimelineMessageDrawerActionType>> = actions
+}
